requirements.txt
faiss==1.9.0
ipython==8.12.3
matplotlib==3.10.0
numpy==2.2.1
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.6.0
scipy==1.14.1
skimage==0.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.66.5
transformers==4.47.1
